Version 1.0
Works on Command line only. Does not error check if there are not photos in specified directory. Images should be the same size or else there will be some cut-off (you want them the same size anyway). Specify a folder on the command line after running, specify a name for the file (The program automatically puts a ".png" at the end of the file), and you will get your sprite sheet as a single image.