cpp sprite sheet generator
SUMMARY:
Added Sprite class which contains image data as well as offsets useful
(possibly?) for the sprite sheet.

Added SpriteSheet class which contains a vector of Sprites and retains
column/row and width/height information for the final sprite sheet.

The main file so far just takes in an input path and an output path as
well as the number of columns. For now, it only prints out the filenames
(with extensions) in that path using batch commands to a file names
temp_dir.txt in the output directory.

NEXT STEPS:
* Need to iterate through the text file generated and load each image into
a sprite that is listed within that text file.
* Need to combine all of those images and print them out as a sprite
  sheet.
* Ensure that all the sprite sheet contains the sprites in the correct
  order.