User can pick from output directory when saving
Before, when saving, the user would have to type out the name of the
file when saving to the output directory.

Now, the user gets a list of files to choose from.

If they select one outside of the index, then they can manually type out
the new file if they wish.